Stores all matrices in new json file format. Updates README.
LDPCStorage.jl now uses a new json-based format. This format is going to be used in this project from now on, as well as the new version of the Julia package.
Generation of C++ header files is now also part of LDPCStorage.jl and the script `ldpc_codegen.jl` in this repository now only implements a command line interface.
Note that the C++ simulator in this project still only supports parsing of binary matrices in the deprecated .cscmat files. This will be fixed later.